requirements.txt
fastapi
uvicorn
openai-whisper
huggingface_hub
pydantic
pyngrok
